mass func separate file
