+++ dymension/libs/kaspa/lib/relayer/src/withdraw/mod.rs
@@ -1,5 +1,6 @@
 pub mod demo;
 pub mod hub_to_kaspa;
+pub mod mass;
 pub mod messages;
 pub mod minimum;
 pub mod sweep;
